removed faculty route and overwrote for-uninversity route
